Convert test_client to async
